requirements.txt
mysql-connector-python==8.0.31
click==8.1.3
pre-commit==2.20.0
